Update README.md
BABYAKITAINU
Baby Akita Inu is another experimental yet revolutionary meme based token, with a burn and is an anti rug, community driven token.
It is early stages here at Baby Akita Inu Token so the main thing is getting the token off the ground and bringing some value.
MOBI is a token built on BSC chain as the key for Babyakitainu. Mobi holders will have the right to trade BABYAKITAINU freely.